Create :dependency-management-base
Split out classes from :core-api that are dependency-management-related and move them into a new project :dependency-management-base.
This new project contains dependency-management-related classes that do not depend on :core or :core-api.
Many things could not be moved that should be moved -- mostly those that depend on domain object containers or the provider API.
We will need to split up model-core before we can move those additional classes.

Some classes like those that depend on Project will have to remain in :core-api and :core until we do further rework.

Many changes to Attributes were required here.
Particularly in the area of serialization.

:model-core and :workers had dependency-management-specific code for serializing Attributes. Work needed
to be done to make serializing of attributes pluggable. We already had some code for pluggable serialization,
but this new work reworked when we create serializers, and as such needed to move some services to different scopes,
particularly in the :workers subproject.

Further, due to initialization order of serializers, and since the AttributesFactory depends on the Isolatable infrastructure,
we needed to split up AttributesFactory into BaseAttributesFactory, the original which depends on serializers for isolating values,
and the base which assumes all values are already isolated and does not depend on pluggable serialization.